Move convert_wkt_to_scene back to fels.py
