+++ fels/fels.py
@@ -11,8 +11,71 @@ from fels.sentinel2 import (
     query_sentinel2_catalogue, get_sentinel2_image, safedir_to_datetime,
     ensure_sentinel2_metadata
 )
-from fels.utils import convert_wkt_to_scene
-import ubelt as ub  # NOQA
+import pkg_resources
+import shapely as shp
+
+
+def convert_wkt_to_scene(sat, geometry, include_overlap):
+    '''
+    Args:
+        sat: 'S2', 'ETM', 'OLI_TIRS'
+        geometry: WKT or GeoJSON string
+        include_overlap: if True, use predicate 'intersects', else use predicate 'contains'
+
+    Returns:
+        List of scenes containing the geometry
+
+    Example:
+        >>> sat = 'S2'
+        >>> geometry = json.dumps({
+        >>>     'type': 'Polygon', 'coordinates': [[
+        >>>         [40.4700, -74.2700],
+        >>>         [41.3100, -74.2700],
+        >>>         [41.3100, -71.7500],
+        >>>         [40.4700, -71.7500],
+        >>>         [40.4700, -74.2700],
+        >>>     ]]})
+        >>> include_overlap = True
+        >>> convert_wkt_to_scene('S2', geometry, include_overlap)
+        >>> convert_wkt_to_scene('LC', geometry, include_overlap)
+    '''
+
+    if sat == 'S2':
+        path = pkg_resources.resource_filename(__name__, os.path.join('data', 'sentinel_2_index_shapefile.shp'))
+    else:
+        path = pkg_resources.resource_filename(__name__, os.path.join('data', 'WRS2_descending.shp'))
+
+    if isinstance(geometry, dict):
+        feat = shp.geometry.shape(geometry)
+    elif isinstance(geometry, str):
+        try:
+            feat = shp.geometry.shape(json.loads(geometry))
+        except json.JSONDecodeError:
+            feat = shp.wkt.loads(geometry)
+    else:
+        raise TypeError(type(geometry))
+
+    # gdf = geopandas.read_file(path)
+    gdf = _memo_geopandas_read(path)
+
+    if include_overlap:
+        # TODO paramatarize thresh
+        thresh = 0.0
+        if thresh > 0:
+            # Requires some minimum overlap
+            overlap = gdf.geometry.intersection(feat).area / feat.area
+            found = gdf[overlap > thresh]
+        else:
+            # Any amount of overlap is ok
+            found = gdf[gdf.geometry.intersects(feat)]
+    else:
+        # This is the bottleneck when the downloaded data exists
+        found = gdf[gdf.geometry.contains(feat)]
+
+    if sat == 'S2':
+        return found.Name.values
+    else:
+        return found.WRSPR.values
 
 
 def normalize_satcode(sat):
@@ -31,13 +94,10 @@ def normalize_satcode(sat):
 
 def get_parser():
     import fels
-    version_info = {
-        'version': fels.__version__,
-    }
+    version_info = {'version': fels.__version__}
 
     parser = argparse.ArgumentParser(
-        prog='fels',
-        description=(
+        prog='fels', description=(
             "Fels {version} - "
             "Find and download Landsat and Sentinel-2 data from the public Google Cloud"
         ).format(**version_info)
@@ -121,6 +181,7 @@ def _get_options(*args, **kwargs):
         >>> options = _get_options(**kwargs)
         >>> print('options = {!r}'.format(options))
         >>> options = _get_options(geometry='POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))')
+        >>> import ubelt as ub
         >>> print('options.__dict__ = {}'.format(ub.repr2(options.__dict__, nl=1)))
     """
     args_names = ['scene', 'sat', 'start_date', 'end_date']
@@ -196,6 +257,7 @@ def _run_fels(options):
         >>>     'list': True,
         >>> }
         >>> options = _get_options(**kwargs)
+        >>> import ubelt as ub
         >>> print('options.__dict__ = {}'.format(ub.repr2(options.__dict__, nl=1)))
         >>> _run_fels(options)
     '''
